Move WorkValidationContext to its own class
