+++ subprojects/execution/src/main/java/org/gradle/internal/execution/UnitOfWork.java
@@ -16,8 +16,6 @@
 
 package org.gradle.internal.execution;
 
-import com.google.common.collect.ImmutableList;
-import com.google.common.collect.ImmutableMultimap;
 import com.google.common.collect.ImmutableSortedMap;
 import org.gradle.api.Describable;
 import org.gradle.api.file.FileCollection;
@@ -28,7 +26,6 @@ import org.gradle.internal.execution.history.changes.InputChangesInternal;
 import org.gradle.internal.execution.workspace.WorkspaceProvider;
 import org.gradle.internal.file.TreeType;
 import org.gradle.internal.fingerprint.CurrentFileCollectionFingerprint;
-import org.gradle.internal.reflect.TypeValidationContext;
 import org.gradle.internal.snapshot.FileSystemSnapshot;
 import org.gradle.internal.snapshot.ValueSnapshot;
 import org.gradle.internal.snapshot.impl.ImplementationSnapshot;
@@ -198,14 +195,6 @@ public interface UnitOfWork extends Describable {
      * Validate the work definition and configuration.
      */
     default void validate(WorkValidationContext validationContext) {}
-
-    interface WorkValidationContext {
-        TypeValidationContext createContextFor(Class<?> type, boolean cacheable);
-
-        ImmutableMultimap<TypeValidationContext.Severity, String> getProblems();
-
-        ImmutableList<Class<?>> getTypes();
-    }
 
     /**
      * Return a reason to disable caching for this work.
